+++ src/parser/parse_prefix.cpp
@@ -5,19 +5,21 @@
 #include "parser/parse_comparation.hpp"
 #include "parser/check.hpp"
 
+#include "ast/expressions/arith.hpp"
+
 expr* parse_prefix(token_list& Token_list) {
 	token* first = Token_list.this_();
 
 	if (check::symbol::is(first, token_symbol::type::PLUS)) {
 		Token_list.next();
 		auto val = parse_multi_divis(Token_list);
-		return new unary_expr { unary_expr::type::positive, val, first->BEGIN, val->END };
+		return new expr_positive { val, first->BEGIN, val->END };
 	}
 
 	else if (check::symbol::is(first, token_symbol::type::MINUS)) {
 		Token_list.next();
 		auto val = parse_multi_divis(Token_list);
-		return new unary_expr { unary_expr::type::negative, val, first->BEGIN, val->END };
+		return new expr_negative { val, first->BEGIN, val->END };
 	}
 
 	else if (check::symbol::is(first, token_symbol::type::NOT)) {
